refactor: 상수 파일 통합 및 구조 개선
- constants.js 파일로 상수 통합 (기존 messages.js 파일 변경)
- 로또 티켓 가격 상수를 constants.js 파일로 이동 (LOTTO_TICKET_PRICE)
- 공통 상수 관리로 코드 일관성과 재사용성 향상
